move get_completion_signatures into __senders.hpp
