+++ include/stdexec/__detail/__senders_core.hpp
@@ -18,8 +18,10 @@
 #include "__execution_fwd.hpp"
 
 // include these after __execution_fwd.hpp
+#include "__awaitable.hpp"
 #include "__completion_signatures.hpp"
 #include "__concepts.hpp"
+#include "__domain.hpp"
 #include "__env.hpp"
 #include "__receivers.hpp"
 #include "__type_traits.hpp"
